Repair dynamic singleton optimization in NetPoller so it understands stacking of multiple Pollers. This occurs frequently, e.g. in stack net.multi.gm or net.bytes.multi.gen.tcp_blk.
From NetPoller in downcall mode, setupConnection now always starts its
subInputs in downcall mode, even if they must be decoupled because this
is no singleton. If it is a singleton, fine. If it is not, the subInput
must be converted to upcall mode (thus introducing a decoupling thread).
This now also works if the subInput is also a Poller.
However, the whole setupConnection mechanism had to be changed to a
2-phase protocol. First all connections along the driver stack are
negotiated, including upcall/downcall and second thoughts about this mode.
If a lowest layer input is switched to upcall, it can immediately start
delivering. In the previous setup, this created a race with the configuration
of the input one layer up. It may still have second thoughts about its
upcall/downcall mode. A separate startReceive call is introduced to start
upcall threads only *after* all negotiations have completed, i.e. after
the uppermost setupConnection has returned.


git-svn-id: http://gforge.cs.vu.nl/svn/ibis/ibis/trunk@2376 aaf88347-d911-0410-b711-e54d386773bb